Move test fixtures to common crate::test::fixtures module
