+++ src/achievement/test_rule.rs
@@ -1,41 +1,5 @@
-use crate::achievement::{process_rules, Achievement, Rule};
-use crate::git::test::fixtures;
-
-pub struct AlwaysFail;
-impl Rule for AlwaysFail {
-    fn name(&self) -> &'static str {
-        "AlwaysFail"
-    }
-    fn process(&mut self, _commit: &git2::Commit, _repo: &git2::Repository) -> Option<Achievement> {
-        None
-    }
-}
-
-pub struct ParticipationTrophy;
-impl Rule for ParticipationTrophy {
-    fn name(&self) -> &'static str {
-        "Participation Trophy"
-    }
-    fn process(&mut self, commit: &git2::Commit, _repo: &git2::Repository) -> Option<Achievement> {
-        tracing::debug!("Granting {:?} a participation trophy", commit.id());
-        Some(Achievement { name: self.name() })
-    }
-}
-
-pub struct ParticipationTrophy2;
-impl Rule for ParticipationTrophy2 {
-    fn name(&self) -> &'static str {
-        "Participation Trophy 2"
-    }
-    fn process(&mut self, _commit: &git2::Commit, _repo: &git2::Repository) -> Option<Achievement> {
-        None
-    }
-
-    fn finalize(&mut self, _repo: &git2::Repository) -> Vec<Achievement> {
-        tracing::debug!("Finalizing ParticipationTrophy2");
-        vec![Achievement { name: self.name() }]
-    }
-}
+use crate::achievement::{process_rules, Rule};
+use crate::test::fixtures;
 
 #[test]
 fn test_no_rules() {
@@ -70,7 +34,7 @@ fn test_iterator_no_matches() {
         }
     });
 
-    let rules = vec![Box::new(AlwaysFail) as Box<dyn Rule>];
+    let rules = vec![Box::new(fixtures::rule::AlwaysFail) as Box<dyn Rule>];
     let achievements = process_rules(oids, &temp_repo.repo, rules);
     let achievements: Vec<_> = achievements.collect();
     assert!(achievements.is_empty());
@@ -90,8 +54,8 @@ fn test_iterator_all_matches() {
     });
 
     let rules = vec![
-        Box::new(AlwaysFail) as Box<dyn Rule>,
-        Box::new(ParticipationTrophy) as Box<dyn Rule>,
+        Box::new(fixtures::rule::AlwaysFail) as Box<dyn Rule>,
+        Box::new(fixtures::rule::ParticipationTrophy) as Box<dyn Rule>,
     ];
     let achievements = process_rules(oids, &temp_repo.repo, rules);
     let achievements: Vec<_> = achievements.collect();
@@ -111,7 +75,7 @@ fn test_awards_on_finalize() {
         }
     });
 
-    let rules = vec![Box::new(ParticipationTrophy2) as Box<dyn Rule>];
+    let rules = vec![Box::new(fixtures::rule::ParticipationTrophy2) as Box<dyn Rule>];
     let achievements = process_rules(oids, &temp_repo.repo, rules);
     let achievements: Vec<_> = achievements.collect();
     assert_eq!(achievements.len(), 1);
