+++ src/git/mod.rs
@@ -1,6 +1,3 @@
-#[cfg(test)]
-pub mod test;
-
 use std::path::Path;
 
 use eyre::WrapErr;
