+++ src/test/fixtures/mod.rs
@@ -1,0 +1,2 @@
+pub mod repository;
+pub mod rule;
