+++ src/test/mod.rs
@@ -1,4 +1,5 @@
 pub mod fixtures;
+
 use tracing::Level;
 use tracing_subscriber::fmt::writer::TestWriter;
 use tracing_subscriber::EnvFilter;
